Dashboard now has container and is presentational
